+++ src/main/java/com/rentcar/BackRentCar/service/AlquilerServicio.java
@@ -2,6 +2,6 @@ package com.rentcar.BackRentCar.service;
 
 import com.rentcar.BackRentCar.model.Alquiler;
 
-public interface AlquilerServicio extends GenericService<Alquiler, Integer> {
+public interface AlquilerServicio extends GenericService<Alquiler, Long> {
 
 }
