+++ src/main/java/com/rentcar/BackRentCar/service/ClaseAutomovilService.java
@@ -2,6 +2,6 @@ package com.rentcar.BackRentCar.service;
 
 import com.rentcar.BackRentCar.model.ClaseAutomovil;
 
-public interface ClaseAutomovilService extends GenericService<ClaseAutomovil, Integer> {
+public interface ClaseAutomovilService extends GenericService<ClaseAutomovil, Long> {
 
 }
