+++ src/main/java/com/rentcar/BackRentCar/service/ComprobanteService.java
@@ -1,5 +1,5 @@
 package com.rentcar.BackRentCar.service;
 import com.rentcar.BackRentCar.model.Comprobante;
-public interface ComprobanteService extends GenericService<Comprobante, Integer> {
+public interface ComprobanteService extends GenericService<Comprobante, Long> {
 
 }
