+++ src/main/java/com/rentcar/BackRentCar/service/PersonaService.java
@@ -2,6 +2,6 @@ package com.rentcar.BackRentCar.service;
 
 import com.rentcar.BackRentCar.model.Persona;
 
-public interface PersonaService extends GenericService<Persona, Integer> {
+public interface PersonaService extends GenericService<Persona, Long> {
 
 }
