+++ src/main/java/com/rentcar/BackRentCar/service/RegistrarDanioService.java
@@ -2,6 +2,6 @@ package com.rentcar.BackRentCar.service;
 
 import com.rentcar.BackRentCar.model.RegistrarDanio;
 
-public interface RegistrarDanioService extends GenericService<RegistrarDanio, Integer> {
+public interface RegistrarDanioService extends GenericService<RegistrarDanio, Long> {
 
 }
