+++ src/main/java/com/rentcar/BackRentCar/service/ReservaService.java
@@ -2,6 +2,6 @@ package com.rentcar.BackRentCar.service;
 
 import com.rentcar.BackRentCar.model.Reserva;
 
-public interface ReservaService extends GenericService<Reserva, Integer> {
+public interface ReservaService extends GenericService<Reserva, Long> {
 
 }
