+++ src/main/java/com/rentcar/BackRentCar/service/RolService.java
@@ -2,6 +2,6 @@ package com.rentcar.BackRentCar.service;
 
 import com.rentcar.BackRentCar.model.Rol;
 
-public interface RolService extends GenericService<Rol, Integer> {
+public interface RolService extends GenericService<Rol, Long> {
 
 }
