+++ src/main/java/com/rentcar/BackRentCar/service/SeguroService.java
@@ -2,7 +2,7 @@ package com.rentcar.BackRentCar.service;
 
 import com.rentcar.BackRentCar.model.Seguro;
 
-public interface SeguroService extends GenericService<Seguro, Integer> {
+public interface SeguroService extends GenericService<Seguro, Long> {
 
 }
